pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "cross-border-agent"
version = "0.1.0"
description = "A multi-agent travel optimizer designed for complex, multi-country, multi-modal trips."
readme = "README.md"
requires-python = ">=3.11"
authors = [ {name = "Daria Tsarova-Lenska"}, { name = "Vladimir Potapov" } ]
license = { text = "MIT" }
dependencies = [
  "google-adk",
  "python-dotenv"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.7",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
